+++ src/api/transaction.ts
@@ -32,13 +32,13 @@ import {
   AnyRawTransaction,
   InputGenerateTransactionOptions,
   InputGenerateTransactionPayloadData,
-  SimpleTransaction,
 } from "../transactions";
 import { AccountAddressInput, Account, PrivateKey } from "../core";
 import { Build } from "./transactionSubmission/build";
 import { Simulate } from "./transactionSubmission/simulate";
 import { Submit } from "./transactionSubmission/submit";
 import { TransactionManagement } from "./transactionSubmission/management";
+import { SimpleTransaction } from "../transactions/instances/simpleTransaction";
 
 export class Transaction {
   readonly config: AptosConfig;