+++ src/api/transactionSubmission/build.ts
@@ -3,12 +3,9 @@
 
 import { AccountAddressInput } from "../../core";
 import { generateTransaction } from "../../internal/transactionSubmission";
-import {
-  InputGenerateTransactionPayloadData,
-  InputGenerateTransactionOptions,
-  SimpleTransaction,
-  MultiAgentTransaction,
-} from "../../transactions";
+import { InputGenerateTransactionPayloadData, InputGenerateTransactionOptions } from "../../transactions";
+import { MultiAgentTransaction } from "../../transactions/instances/multiAgentTransaction";
+import { SimpleTransaction } from "../../transactions/instances/simpleTransaction";
 import { AptosConfig } from "../aptosConfig";
 
 /**