+++ src/transactions/instances/index.ts
@@ -9,3 +9,4 @@ export * from "./rotationProofChallenge";
 export * from "./signedTransaction";
 export * from "./transactionArgument";
 export * from "./transactionPayload";
+export * from "./simpleTransaction";